requirements.txt
adrf==0.1.2
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.2.0
anyio==4.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async-property==0.2.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.2.0
black==23.11.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
cryptography==41.0.7
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.8
django-celery-beat==2.5.0
django-dbbackup==4.0.2
django-redis==5.4.0
django-templated-mail==1.1.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.26.5
exceptiongroup==1.2.0
filelock==3.13.1
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
identify==2.5.33
idna==3.6
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kombu==5.3.4
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.11.2
pgbackup==1.0.1
Pillow==10.1.0
platformdirs==4.1.0
pre-commit==3.5.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.31.1
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.13.2
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.12
yarl==1.9.4
